Refinements in RCTBackedTextInputViewProtocol protocol
Summary:
RCTBackedTextInputViewProtocol is a protocol that describes which capabilities must be exposed for some TextInput-like UIView to be compatible with the rest of TextInput infra. We use that in both implementations. In Classic React Native we use Objective-C runtime mechanisms to call many of those methods (via ViewManager's directives); that masked the problem that the protocol was incomplete.

In Fabric, we call all methods normally, so we need to declare all of them here.

Changelog: [Internal] Fabric-specific internal change.

Reviewed By: sammy-SC

Differential Revision: D18950432

fbshipit-source-id: 3cd5cb57a033f7c43fd5f2c10f9706408e8e6d0c