Configure no-response bot (#28202)
Summary:
Add configuration for [no-response bot](https://probot.github.io/apps/no-response/). This will make it so that issues tagged with the "Needs: Author Feedback" are closed after 21 days of inactivity. More importantly, it will make it so that the label is removed whenever the original author of the issue adds a comment.
## Changelog

[Internal] - N/A
Pull Request resolved: https://github.com/facebook/react-native/pull/28202

Test Plan: Verified no-response is enabled for this repository.

Reviewed By: cpojer

Differential Revision: D20256358

Pulled By: hramos

fbshipit-source-id: 51a11e868081afced9ee26640393f08501a8f103